+++ src/direction.rs
@@ -23,12 +23,6 @@ impl Cardinal {
     }
 }
 
-impl<T: en::Float> Into<Angle<T>> for Cardinal {
-    fn into(self) -> Angle<T> {
-        self.angle()
-    }
-}
-
 impl Neg for Cardinal {
     type Output = Self;
 
@@ -91,12 +85,6 @@ impl Neg for Direction {
             West => East,
             Northwest => Southeast,
         }
-    }
-}
-
-impl<T: en::Float> Into<Angle<T>> for Direction {
-    fn into(self) -> Angle<T> {
-        self.angle()
     }
 }